// src/generated/prisma/enums.ts
export const EventStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  UNDER_REVIEW: 'UNDER_REVIEW',
  CORRECTION_REQUIRED: 'CORRECTION_REQUIRED',
  APPROVED: 'APPROVED',
  ARCHIVED: 'ARCHIVED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]
// ...
export const FundingSource = {
  CLUB_FUND: 'CLUB_FUND',
  ROTARY_CLUB: 'ROTARY_CLUB',
  SPONSOR: 'SPONSOR',
  DONATION: 'DONATION',
  MEMBER_CONTRIBUTION: 'MEMBER_CONTRIBUTION',
  PARTNER_ORGANIZATION: 'PARTNER_ORGANIZATION',
  OTHER: 'OTHER'
} as const

export type FundingSource = (typeof FundingSource)[keyof typeof FundingSource]


export const BeneficiaryCategory = {
  CHILDREN: 'CHILDREN',
  STUDENTS: 'STUDENTS',
  WOMEN: 'WOMEN',
  MEN: 'MEN',
  SENIOR_CITIZENS: 'SENIOR_CITIZENS',
  PATIENTS: 'PATIENTS',
  FAMILIES: 'FAMILIES',
  SCHOOL_STUDENTS: 'SCHOOL_STUDENTS',
  COLLEGE_STUDENTS: 'COLLEGE_STUDENTS',
  DIFFERENTLY_ABLED: 'DIFFERENTLY_ABLED',
  GENERAL_PUBLIC: 'GENERAL_PUBLIC',
  COMMUNITY: 'COMMUNITY',
  ENVIRONMENT: 'ENVIRONMENT',
  ANIMALS: 'ANIMALS',
  ROTARACTORS: 'ROTARACTORS',
  OTHER: 'OTHER'
} as const

export type BeneficiaryCategory = (typeof BeneficiaryCategory)[keyof typeof BeneficiaryCategory]